Try removing fso table altogether
OK, you may argue that the database is not completely
normalized now, but it's actually hard to say. In practice,
I think this is actually less redundant.

The fso table contained (id, path, digest) where id was a function
of (path, digest). The id seemed superfluous, so I removed it.
And then, any other table that needed to point to a fso needed
BOTH the path and the digest, i.e. the whole content of the fso row.