Simplify config; Add default config
Now the config is essentially a simple key-value storage,
but it allows complex (YAML) values. The CLI interprets all user
input values as YAML and crashes if it doesn't work.

The immediate benefit with this is that we now can have
complex values (e.g. logging config) in the same place.