Make tasks more reusable, composable
Syntax-wise, the innovation is that a Definition is created by
*calling* a Task with a set of other Definitions as arguments.

Thus, the Task is really independent of the Definition, and may easily
be used to create different Definitions.

This also means that anyone can build a library of common tasks.

The idea here is also to get rid of the different Instruments and instead
make the handling of e.g. JSON data more explicit. This does not have to be
inconvenient, if there is a decent library of functions to make common
JSON transformation/loading/etc operations.

Not sure if the different Instruments idea is orthogonal to this syntax
or not. But it just seems simpler to have all resources represented by a path.