Use tuples of inputs to Calc and Comp, not dicts
Ironically this is very similar to what I changed away from
just recently. Now I remembered why it was constructed this way
in the first place.

The practical result is that a Loc is tied to a Comp.
No need and no possibility to implicitly define a name change of
an output when it is used as input elsewhere.

The change of Calc inputs to a tuple makes no semantical difference
but I'm doing it anyway for consistency between Calc and Comp.

Renaming comp_input -> parent and calc_input -> input like it was before.

This imposes no limitation on what the system can do, but
it does require an explicit move/rename of an output if that is
desired.

The change is good for these reasons:

 * Explicit is better than implicit.
 * Reduces mental strain that arises due to the possibly different names
   of inputs and outputs. Now the Comp can be seen as a definition of a file
   at a given location until it is explicitly renamed.
 * Drops the dependency on the immutable map; now the Comps and Calcs
   are trivial to hash since the input collections are tuples.