Create new database schema with indexed nodes
Nodes (previously Comps) optionally have an index. Nodes without index
represent a resource (think a file or directory); nodes with index
represent an array of resources (think a directory of files or subdirs).

The index is abstractly represented by a Defn. This is because the
concrete index may not be known when the graph is defined: for example,
the concrete index may depend on a glob operation in a directory
which is the output of an upstream Defn.

- - -

The previous "comp" table is now replaced by tables "node" and "defn".

Node is maybe a better name as it more obviously suggests a graph.

A Defn is an output of a Node, i.e., the Defn represents a definition
of an output, while the Node rather represents the calculation step
that will produce the output (and possibly siblings).