Distinguish Calc from CalcBundle
- A NodeBundle corresponds to a CalcBundle
- A Node corresponds to a Calc

The Node and the Calc are the simplest units of calculation, and they
are what we should put into cache and provenance databases. But
in practice a CalcBundle can be run in one go.

It should be noted that NodeBundle and CalcBundle
do not define output trees, but rather a Mapping[Path, TreeItem],
which is the collection of results of their individual Node/Calc
components. The TreeItem is the result unit saved in the cache and
provenance databases.